requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0